code clean & architecture
